waveplates_tracker/requirements.txt
numpy
pillow
pystray
pyautogui
pywin32
opencv-python
pytesseract